requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.7
django-cors-headers==4.3.1
django-cte==1.3.2
django-debug-toolbar==4.2.0
django-filter==23.4
django-import-export==3.3.3
django-nested-inline==0.4.6
django-redis==5.4.0
django-resized==1.0.2
django-rich-logging==0.3.0
django-richtextfield==1.6.1
django-simple-history==3.4.0
django-templated-mail==1.1.1
django-user-agents==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
dnspython==2.4.2
drf-yasg==1.21.7
email-validator==2.1.0.post1
et-xmlfile==1.1.0
gando==0.7.8
gunicorn==21.2.0
h11==0.14.0
idna==3.6
inflection==0.5.1
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupPy==1.14
mdurl==0.1.2
numpy==1.26.2
oauthlib==3.2.2
odfpy==1.4.1
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
Pygments==2.17.2
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.8.0
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.1.0
user-agents==2.2.0
uvicorn==0.24.0.post1
xlrd==2.0.1
xlwt==1.3.0
